+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added emails watcher ([c19b9b0](https://github.com/ydethe/ragindexer/commit/c19b9b040521dafa0c71219bb1bcf220d0959877) by Yann de The).
 - Added emails observer ([511da4a](https://github.com/ydethe/ragindexer/commit/511da4ad52cce5a7c21ebe0b26cc961658565be2) by Yann de The).
 - Added API keys and remove of deprecated method QdrantClient.search ([12d727f](https://github.com/ydethe/ragindexer/commit/12d727f8c96d018c1ca1d803222664a7ca528fca) by Yann de The).
 - Added page progress ([fd78e72](https://github.com/ydethe/ragindexer/commit/fd78e7266a6019c9934b22e044b614420da79e9c) by Yann de The).